chore(examples): update migration example
