+++ examples/demos/migrate/nodejs/src/utils.ts
@@ -1,0 +1,15 @@
+import readline from "readline";
+
+export const askQuestion = (query: string): Promise<string> => {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+
+  return new Promise((resolve) => {
+    rl.question(query, (answer) => {
+      rl.close();
+      resolve(answer);
+    });
+  });
+};
